Move trait to `store` modules, and figure out Send strategy
